code updates - splitting formatters for convenience and group peers by ns for dot
