+++ pkg/netpol/connlist/conns_formatter_md.go
@@ -1,0 +1,33 @@
+package connlist
+
+import (
+	"fmt"
+	"sort"
+	"strings"
+)
+
+// formatMD: implements the connsFormatter interface for md output format
+type formatMD struct {
+}
+
+// formats the md output header
+func getMDHeader() string {
+	return "| src | dst | conn |\n|-----|-----|------|"
+}
+
+// formats a connection line for md output
+func getMDLine(c singleConnFields) string {
+	return fmt.Sprintf("| %s | %s | %s |", c.Src, c.Dst, c.ConnString)
+}
+
+// returns a md string form of connections from list of Peer2PeerConnection objects
+func (md formatMD) writeOutput(conns []Peer2PeerConnection) (string, error) {
+	mdLines := make([]string, len(conns))
+	for index := range conns {
+		mdLines[index] = getMDLine(formSingleConn(conns[index]))
+	}
+	sort.Strings(mdLines)
+	allLines := []string{getMDHeader()}
+	allLines = append(allLines, mdLines...)
+	return strings.Join(allLines, newLineChar), nil
+}
